Added support for floats

+++ quicksort.py
@@ -8,7 +8,7 @@ def quicksort(arr):
 
 def main():
     parser = argparse.ArgumentParser(description='Sort an array of integers using quicksort.')
-    parser.add_argument('integers', metavar='Numbers', type=int, nargs='+',
+    parser.add_argument('integers', metavar='Numbers', type=float, nargs='+',
                         help='The list of integers to sort')
 
     args = parser.parse_args()
